Implement subtler rules for emphasis near punctuation in C parser.
Background:
http://talk.commonmark.org/t/emphasis-inside-strong-broken-in-js-implementation-when-parenthesis-involved/903/6

Still need to update spec and js parser.

The basic idea of the change is that if the delimiter has punctuation
to the left and a normal character (non-space, non-punctuation) to
the right, it can only be an opener.  If it has punctuation to the
right and a normal character (non-space, non-punctuation) to the left,
it can only be a closer.

This handles cases like

    **Gomphocarpus (*Gomphocarpus physocarpus*, syn. *Asclepias
    physocarpa*)**

and

    **foo "*bar*" foo**

better than before.